% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation-data.R
\docType{data}
\name{annotation}
\alias{annotation}
\title{Annotation of diagnosis/procedure code.}
\format{
A data frame with 12,846 rows and 2 columns:
\describe{
\item{code}{ICD-10 code of a diagnosis/procedure (2016). This connects to
\code{diagnosis} data.}
\item{desc}{A character describing diagnosis/procedure.}
}
}
\source{
Artificial data
}
\usage{
annotation
}
\description{
A dataset containing the annotation of diagnosis/procedure code.
}
\keyword{datasets}
