% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis-data.R
\docType{data}
\name{diagnosis}
\alias{diagnosis}
\title{Diagnosis and procedure.}
\format{
A data frame with 4,663,453 rows and 3 columns:
\describe{
\item{visit_id}{Visit ID. Each ID is unique to one episode (from
admission to discharge) per provider. This connects to
\code{visit_cap}, \code{visit_ffs}, and \code{visit_drg} data.}
\item{code}{ICD-10 code of a diagnosis/procedure (2016). This connects to
\code{annotation} data.}
\item{code_type}{Categorical variable of code type, consisting: (1)
\code{Admission diagnosis}; (2) \code{Discharge, primary diagnosis};
(3) \code{Discharge, secondary diagnosis}; and (4) \code{Procedure}.}
}
}
\source{
Artificial data
}
\usage{
diagnosis
}
\description{
A dataset containing the diagnoses and procedures (if any) Every visit is
encountered at least a code of discharge, primary diagnosis. But, a visit may
be encountered 1 or more other code of diagnoses/procedures.
}
\keyword{datasets}
