% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edata-data.R
\docType{data}
\name{edata}
\alias{edata}
\title{Experimenter information for medical history datasets with specific outcome.}
\format{
A MIAME object with 6 attributes:
\describe{
\item{name}{A character of the experimenter name.}
\item{lab}{A character of the laboratory name.}
\item{contact}{A character of the experimenter email.}
\item{title}{A character of the dataset title.}
\item{abstract}{A character of the dataset abstract.}
\item{url}{A character of the dataset URL.}
}
}
\source{
Artificial data
}
\usage{
edata
}
\description{
Experimenter information for experiment data table.
}
\keyword{datasets}
