% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_nps_mh-function.R
\name{extract_nps_mh}
\alias{extract_nps_mh}
\title{Extract medical history without perfect separation}
\usage{
extract_nps_mh(tidy_set, sd_cutoff = 0)
}
\arguments{
\item{tidy_set}{A TidySet (i.e. ExpressionSet) containing the visits of subjects in
outcome dataset, paid by any payment systems. This TidySet also
accomodates outcome dataset. This is an output of
\code{compile_mh_outcome()}.}

\item{sd_cutoff}{Cutoff of standard deviation, a non-negative numeric that is
used for cutoff. If standard deviation of a medical history metric for an
outcome is more than the cutoff for that outcome, either non-event or event,
then that medical history is accounted to the list.}
}
\value{
Medical history list, a data frame with rows for unique medical
history and columns of \code{key}, \code{nonevent} and \code{event}. The last
two columns are standard deviation of the medical history for non-event and
event. Missing value or NA is not accounted when computing the standard
deviation.
}
\description{
This function makes a list of medical histories without perfect separation.
This means a medical history per outcome have non-zero variance.
}
\examples{

## Create input example
data(medhistdata)

## Extract medical history without perfect separation
ps_remover=extract_nps_mh(medhistdata)
}
\keyword{history,}
\keyword{medical}
\keyword{perfect}
\keyword{separation}
