% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_nzv_mh-function.R
\name{extract_nzv_mh}
\alias{extract_nzv_mh}
\title{Extract medical history without zero variance}
\usage{
extract_nzv_mh(tidy_set, sd_cutoff = 0)
}
\arguments{
\item{tidy_set}{A TidySet (i.e. ExpressionSet) containing the visits of subjects in
outcome dataset, paid by any payment systems. This TidySet also
accomodates outcome dataset. This is an output of
\code{compile_mh_outcome()}.}

\item{sd_cutoff}{Cutoff of standard deviation, a non-negative numeric that is
used for cutoff. If standard deviation of a medical history metric is more
than the cutoff, then that medical history is accounted to the list.}
}
\value{
Medical history list, a data frame with rows for unique medical
history and columns of \code{key} and \code{sd_value}. Missing value or NA
is not accounted when computing the standard deviation.
}
\description{
This function makes a list of medical histories without zero variance.
}
\examples{

## Create input example
data(medhistdata)

## Extract medical history without zero variance
zv_remover=extract_nzv_mh(medhistdata)
}
\keyword{history,}
\keyword{medical}
\keyword{variance}
\keyword{zero}
