% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdata-data.R
\docType{data}
\name{mdata}
\alias{mdata}
\title{Metadata for medical history datasets with specific outcome.}
\format{
A data frame with 7 rows and 1 column:
\describe{
\item{labelDescription}{A character to describe the variables.}
}
}
\source{
Artificial data
}
\usage{
mdata
}
\description{
Metadata for phenotype and protocol tables.
}
\keyword{datasets}
