% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medhistdata-data.R
\docType{data}
\name{medhistdata}
\alias{medhistdata}
\title{Medical history datasets with specific outcome.}
\format{
A data frame with 3,434 rows and 7 columns:
\describe{
\item{exprs}{ICD-10 code of a diagnosis/procedure (2016). This connects to
\code{pData()} by columns (visits) and \code{fData()} by rows (medical
histories).}
\item{outcome}{A factor of which non-event is the first class between
non-event and event. Subjects with events are those encountered by
either O14 or O15 while those with non-events encountered by Z33 to
Z37. Called by \code{pData()}.}
\item{desc}{A character describing diagnosis/procedure. Called by
\code{fData()}.}
\item{visit_id}{Visit ID. Each ID is unique to one episode (from admission
to discharge) per provider. Called by \code{pData(protocolData())}.}
\item{subject_id}{Subject ID. Each ID is unique to one subject across
healthcare providers. Called by \code{pData(protocolData())}.}
\item{latest_date}{Latest date of event/non-event. For event, this is
admission_date of subjects at 2 days before the date of event, which is
earliest date of either O14 or O15 encounter. For non-event, this is
admission_date of subjects at the date of non-event, which is the
latest date of Z33 to Z37 encounters. Called by
\code{pData(protocolData())}.}
\item{healthcare_id}{Provider ID. The provider is the one of which the subject visits
(not always one the subject registered to). This connects to
healthcare data. Called by \code{pData(protocolData())}.}
\item{admission_date}{Admission date of this visit. Called by
\code{pData(protocolData())}.
\item{db_start_date}{Start date of the database. Called by
\code{pData(protocolData())}.}
}
}
\source{
Artificial data
}
\usage{
medhistdata
}
\description{
A TidySet (i.e. ExpressionSet) containing the visits of subjects in
outcome dataset, paid by any payment systems. This TidySet also
accomodates outcome dataset.
}
\keyword{datasets}
