% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_binary-function.R
\name{trans_binary}
\alias{trans_binary}
\title{Transform medical history by binarizing day interval}
\usage{
trans_binary(tidy_set, verbose = T)
}
\arguments{
\item{tidy_set}{A TidySet (i.e. ExpressionSet) containing the visits of
subjects in outcome dataset, paid by any payment systems. This TidySet also
accomodates outcome dataset. This is an output of
\code{compile_mh_outcome()}.}

\item{verbose}{Verbosity, a logical indicating whether progress should be
shown.}
}
\value{
A TidySet (i.e. ExpressionSet) containing the visits of subjects in
medical history and the outcome datasets. Since a code may be encountered
before \code{db_start_date}, causing the value returns NA to denote censored
data, that medical history is transformed to 0. But, if the interval is 0,
the value returns 1 because this means the record exists.
}
\description{
This function transforms each medical history from day interval to binary
one. If there is no record, then the value is 0; otherwise, the value is 1.
}
\examples{

## Create input example
data(medhistdata)

## Transform medical history by binarizing day interval
mh_binary=trans_binary(medhistdata)
}
\keyword{binarization}
\keyword{day}
\keyword{history,}
\keyword{interval}
\keyword{medical}
\keyword{of}
