% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visit_drg-data.R
\docType{data}
\name{visit_drg}
\alias{visit_drg}
\title{Visits paid by diagnosis-related group payment system.}
\format{
A data frame with 906,905 rows and 8 columns:
\describe{
\item{visit_id}{Visit ID. Each ID is unique to one episode (from
admission to discharge) per provider. This connects to
\code{diagnosis} data.}
\item{subject_id}{Subject ID. This connects to \code{subject} data.}
\item{healthcare_id}{Provider ID. The provider is the one of which the
subject visits (not always one the subject registered to). This
connects to \code{healthcare} data.}
\item{admission_date}{Admission date of this visit.}
\item{discharge_date}{Discharge date of this visit.}
\item{discharge_status}{Categorical variable of discharge status, consisting
\code{discharged to home}, \code{transferred to other facility},
\code{transferred to higher-level care},
\code{expired}, and \code{expired, place unknown}.}
\item{claimed_cost}{Amount of cost based on claimed diagnosis-related
group, in a cost unit.}
\item{verified_cost}{Amount of cost based on verified diagnosis-related
group, in a cost unit.}
}
}
\source{
Artificial data
}
\usage{
visit_drg
}
\description{
A dataset containing the visits. All visits of a subject sample are included
of which admission dates range from January 1st, 2015 to December 31st, 2016.
The discharge date may be up to 2017. In this table, only visits from
secondary/tertiary care are included. These visits are paid based on
diagnosis-related group. Therefore, there is claimed or verified cost.
}
\keyword{datasets}
